Do not rebuild the app when launching simulation (#490)
In the top Makefile, `make run-app-verilator` has a dependency on `make app`,
and `make app` has a dependency on `make clean-app`.
As a result, running `make app` always rebuilds the app (which is nice),
but running `make run-app-verilator` will also rebuild it, even if it was already built
(with specific parameters, e.g. `PROJECT=not_hello_world`).
This was reported in issue #490.

This commit preserves the `app: clean-app` dependency,
but doesn't rebuild the app on `make run-app-verilator`.